Refactor: 세번째 PR 피드백 리팩토링
