+++ app/src/components/Modal/Carousel.jsx
@@ -1,23 +1,5 @@
 import { useEffect, useRef, useState } from "react";
-import {
-  Container,
-  Header,
-  Title,
-  Counter,
-  LeftBtn,
-  RightBtn,
-  CurrentPage,
-  DividePage,
-  TotalPage,
-  CardCarousel,
-  CardCarouselInner,
-  Card,
-  CardImg,
-  CardTitle,
-  CardPrices,
-  SPrice,
-  NPrice,
-} from "../../styles/Modal/Carousel.js";
+import styled from "styled-components";
 
 const Carousel = () => {
   const CARD_VISIBLE = 5;
@@ -46,10 +28,9 @@ const Carousel = () => {
   }, []);
 
   const moveCarouselPos = () => {
-    // eslint-disable-next-line prettier/prettier
-    carouselRef.current.style.transform = `translateX(${(CARD_SIZE + CARD_GAP) * carouselPosRef.current * -1
-      // eslint-disable-next-line prettier/prettier
-      }px)`;
+    carouselRef.current.style.transform = `translateX(${
+      (CARD_SIZE + CARD_GAP) * carouselPosRef.current * -1
+    }px)`;
   };
 
   const checkActiveBtn = (count) => {
@@ -57,7 +38,7 @@ const Carousel = () => {
     setCurrentCarouselCount(count);
   };
 
-  const handleLeftBtnClick = () => {
+  const handleClickLeftBtn = () => {
     if (carouselBtnRef.current && currentCarouselCount > 1) {
       if (carouselBeforePosRef.current !== CARD_VISIBLE)
         carouselPosRef.current -= carouselBeforePosRef.current;
@@ -70,7 +51,7 @@ const Carousel = () => {
     }
   };
 
-  const handleRightBtnClick = () => {
+  const handleClickRightBtn = () => {
     if (carouselBtnRef.current && currentCarouselCount < totalCarouselCount) {
       const cardRemain = cardCount - CARD_VISIBLE - carouselPosRef.current;
 
@@ -87,7 +68,7 @@ const Carousel = () => {
     }
   };
 
-  const handleBtnTransitionEnd = () => {
+  const handleTransitionEndBtn = () => {
     carouselBtnRef.current = true;
   };
 
@@ -99,28 +80,30 @@ const Carousel = () => {
             <Title>함께하면 더욱 맛있는 상품</Title>
             <Counter>
               <LeftBtn
-                // eslint-disable-next-line prettier/prettier
-                src={`${process.env.PUBLIC_URL}/images/${currentCarouselCount === 1 ? "ModalCarouselLeftDisable.svg" : "ModalCarouselLeft.svg"
-                  // eslint-disable-next-line prettier/prettier
-                  }`}
-                onClick={handleLeftBtnClick}
+                src={`${process.env.PUBLIC_URL}/images/${
+                  currentCarouselCount === 1
+                    ? "ModalCarouselLeftDisable.svg"
+                    : "ModalCarouselLeft.svg"
+                }`}
+                onClick={handleClickLeftBtn}
               ></LeftBtn>
               <CurrentPage>{currentCarouselCount}</CurrentPage>
               <DividePage>/</DividePage>
               <TotalPage>{totalCarouselCount}</TotalPage>
               <RightBtn
-                // eslint-disable-next-line prettier/prettier
-                src={`${process.env.PUBLIC_URL}/images/${currentCarouselCount === totalCarouselCount ? "ModalCarouselRightDisable.svg" : "ModalCarouselRight.svg"
-                  // eslint-disable-next-line prettier/prettier
-                  }`}
-                onClick={handleRightBtnClick}
+                src={`${process.env.PUBLIC_URL}/images/${
+                  currentCarouselCount === totalCarouselCount
+                    ? "ModalCarouselRightDisable.svg"
+                    : "ModalCarouselRight.svg"
+                }`}
+                onClick={handleClickRightBtn}
               ></RightBtn>
             </Counter>
           </Header>
           <CardCarousel>
             <CardCarouselInner
               ref={carouselRef}
-              onTransitionEnd={handleBtnTransitionEnd}
+              onTransitionEnd={handleTransitionEndBtn}
               gap={CARD_GAP}
             >
               {carouselDish.map((e) => {
@@ -148,5 +131,107 @@ const Carousel = () => {
     </>
   );
 };
+
+const Container = styled.div`
+  padding: 4.8rem 4.8rem 0 4.8rem;
+`;
+
+const Header = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
+
+const Title = styled.h3`
+  font-weight: ${({ theme }) => theme.fontWeight.bold};
+  font-size: ${({ theme }) => theme.fontSize.large};
+`;
+
+const Counter = styled.div`
+  display: flex;
+  gap: 0.8rem;
+`;
+
+const Btn = styled.button`
+  background-repeat: no-repeat;
+  background-position: center;
+  width: 2.4rem;
+  border: none;
+  background-color: ${({ theme }) => theme.color.white};
+  cursor: pointer;
+`;
+
+const LeftBtn = styled(Btn)`
+  background-image: url(${({ src }) => src});
+`;
+
+const RightBtn = styled(Btn)`
+  background-image: url(${({ src }) => src});
+`;
+
+const PageText = styled.div`
+  font-size: ${({ theme }) => theme.fontSize.small};
+  line-height: 2.4rem;
+`;
+
+const CurrentPage = styled(PageText)`
+  font-weight: ${({ theme }) => theme.fontWeight.bold};
+`;
+
+const DividePage = styled(PageText)`
+  font-weight: ${({ theme }) => theme.fontWeight.regular};
+`;
+
+const TotalPage = styled(DividePage)`
+  font-weight: ${({ theme }) => theme.fontWeight.regular};
+`;
+
+const CardCarousel = styled.div`
+  overflow: hidden;
+  width: 86.4rem;
+  height: 23.6rem;
+  margin-top: 2.8rem;
+`;
+
+const CardCarouselInner = styled.div`
+  display: flex;
+  flex-wrap: nowrap;
+  gap: ${({ gap }) => `${gap / 10}rem`};
+  transition: all 1s;
+`;
+
+const Card = styled.div`
+  flex-shrink: 0;
+  flex-basis: ${({ size }) => `${size / 10}rem`};
+  cursor: pointer;
+`;
+
+const CardImg = styled.img`
+  width: 100%;
+`;
+
+const CardTitle = styled.h3`
+  margin-top: 0.8rem;
+  line-height: 2.4rem;
+  font-weight: ${({ theme }) => theme.fontWeight.regular};
+  font-size: ${({ theme }) => theme.fontSize.small};
+  color: ${({ theme }) => theme.color.grey1};
+`;
+
+const CardPrices = styled.div`
+  display: flex;
+  gap: 0.8rem;
+`;
+
+const SPrice = styled.div`
+  font-size: ${({ theme }) => theme.fontSize.medium};
+  font-weight: ${({ theme }) => theme.fontWeight.bold};
+`;
+
+const NPrice = styled.div`
+  font-size: ${({ theme }) => theme.fontSize.small};
+  font-weight: ${({ theme }) => theme.fontWeight.regular};
+  color: ${({ theme }) => theme.color.grey3};
+  text-decoration: line-through;
+`;
 
 export default Carousel;
