+++ app/src/index.jsx
@@ -3,8 +3,8 @@ import ReactDOM from "react-dom/client";
 import App from "./components/App";
 import { ThemeProvider } from "styled-components";
 import { Reset } from "styled-reset";
-import GlobalStyle from "./styles/global.js";
-import theme from "./styles/theme.js";
+import GlobalStyle from "./styled-components/global.js";
+import theme from "./styled-components/theme.js";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
